requirements.txt
alembic==1.15.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3